Added admin functionality

+++ app/controllers/chatroom_users_controller.rb
@@ -4,12 +4,19 @@ class ChatroomUsersController < ApplicationController
   include CurrentChatroom
 
   def destroy
-    current_chatroom=Chatroom.find(params[:chatroom])
+    current_chatroom = Chatroom.find(params[:chatroom])
     @chatroom_user = ChatroomUser.find_by(chatroom_id: params[:chatroom], user_id: params[:user])
+    admin = @chatroom_user.admin
     @chatroom_user.destroy
-    redirect_to chatrooms_path
-    if current_chatroom.users.count == 0
-      current_chatroom.destroy
+    current_chatroom.destroy if current_chatroom.users.count == 0
+
+    if params[:user] == current_user.id.to_s
+      redirect_to chatrooms_path
+    end
+
+    if admin && current_chatroom.users.count > 0
+      next_user = current_chatroom.chatroom_users.first
+      next_user.update(admin: true)
     end
   end
 end
